FileDigestCheck: move digester types into utility module
This helps keeping the main check implementation and utility helpers
better apart from each other.
